default the linux_job workflow to pull submodules recursively
Summary:
When trying to use this for the "codegen diff" job in PyTorch, I
noticed it was failing because we did not have submodules
available. After discussing with @seemethere, he suggested to just
make this the default.

Test Plan:
Verifying in local CI and testing with the codegen diff job before
merging.